feat: add initial `esphome` config for `ESP8266`
This is and always will be a counterpart version
of the `ESP32` config.

This config was compiled into a firmware and successfully flashed,
however the PWM control output itself needs to be tested yet.
But it's expected to work given that the config is almost same
as the `ESP32` one.